// drawsysByVscode/src/type/type.ts
export interface TreeNodeType<T> {
  treeId: string;
  treeLabel: string;
  expanded: boolean;
  parentId: string | number;
  level: number;
  source: T;
  children?: TreeNodeType<T>[]; // 在 buildTree 之后才会赋值
  shouldRename?: boolean;
}
// ...
export interface StructAndVarRelationTreeNodeBase {
  id: string;
  name: string;
  comment: string;
  dataType: string;
  length: string;
  formatAdjustment: string;
  connection: string;
  group: string;
  address: string;
  linearCalibration: string;
  asRangeFrom: string;
  asRangeTo: string;
  osRangeFrom: string;
  osRangeTo: string;
  lowerLimit: string;
  upperLimit: string;
  iconType?: string;
  type: "folder" | "file"; // 限制类型为文件夹或文件
  connectionParameters: string;
  children?: StructAndVarRelationTreeNodeBase[]; // 子节点（仅文件夹有）
  agreementId?: string;
  agreement?: Object;
  // [key: string]: any;
}
// ...
export interface optionType {
  label: string | number | Function;
  value: string | number | Function;
  [key: string]: any;
}